requirements.txt
streamlit
python-dotenv
openai
langchain
faiss-cpu
pymupdf
langchain-community